requirements.txt
# DRfold2 Training Requirements

# Core dependencies
torch>=1.11.0
torchvision>=0.12.0
torchaudio>=0.11.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Bioinformatics
biopython>=1.79

# Configuration and logging
pyyaml>=5.4.1
tensorboard>=2.8.0
tqdm>=4.62.0

# Data processing
pandas>=1.3.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0

# Development (optional)
ipython>=8.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
